for Windows: don't attempty to kill the Logisim process because of permission denied error

+++ tests/sanity_test.py
@@ -38,7 +38,9 @@ class TestCase():
       debug_buffer = [] 
       passed = compare_unbounded(proc.stdout,reference,debug_buffer)
     finally:
-      os.kill(proc.pid,signal.SIGTERM)
+      if os.name != 'nt':
+        os.kill(proc.pid,signal.SIGTERM)
+
     if passed:
       return (True, "Matched expected output")
     else:
